\name{grab.cxxfun-methods}
\docType{methods}
\alias{grab.cxxfun}
\alias{grab.cxxfun-methods}
\alias{grab.cxxfun,cxxdso-method}
\title{Retrieve the functions object associated with 
       an object of class \code{cxxdso}} 
\description{
  The \code{grab.cxxfun} function retrieves the function object
  associated with objects of class \code{cxxdso} generated by
  \code{\link{cxxfunctionplus}}
}
\section{Methods}{
\describe{

\item{\code{signature(x = "cxxdso")}}{Retrieves the function object
  associated with the \code{cxxdso} objects generated by
  \code{\link{cxxfunctionplus}}.}  
}}

\seealso{
\code{cxxfunctionplus}, \code{\linkS4class{cxxdso}} } 
\examples{
\dontrun{
dso <- cxxfunctionplus(signature(), "return R_NilValue;")
dll <- grab.cxxfun(dso)
}} 

